Rename FS queue package to internal
