+++ federationsender/internal/api.go
@@ -1,4 +1,4 @@
-package query
+package internal
 
 import (
 	"encoding/json"
