+++ federationsender/internal/query.go
@@ -1,4 +1,4 @@
-package query
+package internal
 
 import (
 	"context"
